requirements.txt
openai-agents
openai
anthropic
python-dotenv
fastapi[standard]
python-zalo-bot
litellm[proxy]
logfire